to deploy

+++ jobBackend/jobsControllers/deleteJobController.js
@@ -7,16 +7,16 @@ const deleteJobController = async (req, res, next) => {
     try {
         // get the postId from the params
         const { postId } = req.params;
-        // return no post to like
+        // return no post to delete
         if (!postId){
-            return res.status(200).json({ message: "No post selected for deletion" });
+            return res.status(400).json({ message: "No post selected for deletion" });
         }
         // get the id of user from the verifiedToken of user in cookie
-        // const userId = req.userId;
+        const userId = req.userId;
         // throw error if no user Id
-        // if (!userId) {
-            // throw new CustomError("You have to login first", 401);
-        // }
+        if (!userId) {
+            throw new CustomError("You have to login first", 401);
+        }
         // get post object
         const postToDelete = await Job.findById(postId);
         // throw error if not found
@@ -24,31 +24,23 @@ const deleteJobController = async (req, res, next) => {
             throw new CustomError("Post not found", 404);
         }
         // get user that posted
-        // const user = await User.findById(postToDelete.user);
+        const user = await User.findById(userId);
         // throw error if not found
-        // if (!user) {
-            // throw new CustomError("User not found", 404);
-        // }
-        // check if the post owner is logged user
-        // if (postToDelete.user.toString() !== userId.toString()){
-            // return res.status(401).json("You are not authorized to delete this post");
-        // }
-        // Remove the postId from the user's list of posts
-        // user.posts=user.posts.filter(postId=>postId.toString()!==postToDelete._id.toString());
-        // save the user 
-        // await user.save();
+        if (!user) {
+            throw new CustomError("User not found", 404);
+        }
+        // check if the user is admin or the post owner
+        if (user.role !== 'admin') {
+            return res.status(403).json("You are not forbidden to delete this post");
+        }
         // Delete the post by its ID
         await Job.findByIdAndDelete(postId);
-        // Define cache key
-        // const cacheKey = `post_${userId}`;
-        // const cacheKey1 = `userpost_${userId}`;
-        // clear cache of the user
         // clearCache(cacheKey);
         // clearCache(cacheKey1);
 
         res.status(200).json({message:"Post deleted successfully"});
     } catch (error) {
-        next (error);
+        next(error);
     }
 }
